pyproject.toml
[project]
name = "carservicevoiceassistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.8",
    "livekit-agents[openai,silero,turn-detector]~=1.4",
    "livekit-api>=1.1.0",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["carservicevoiceassistant"]
